experiments/pytorch-ignite-hydra-template/requirements.txt
# --------- basics --------- #
numpy==1.26.4
scipy==1.12.0

# --------- pytorch --------- #
torch==2.2.0
torchvision==0.17.0
pytorch-ignite==0.4.13

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
omegaconf==2.3.0

# --------- loggers --------- #
tqdm==4.66.2
tensorboardX==2.6.2.2
tensorboard==2.16.2
